well formated json
can't post nor see relations

+++ db/migrate/20170523122626_create_datacentres.rb
@@ -16,7 +16,7 @@ class CreateDatacentres < ActiveRecord::Migration[5.1]
       t.string :symbol
       t.datetime :updated
       t.integer :version
-      t.integer :allocator
+      t.bigint :allocator
       t.string :experiments
 
       t.timestamps
